+++ clash-protocols/tests/Tests/Protocols/Avalon.hs
@@ -23,9 +23,9 @@ import Test.Tasty.TH (testGroupGenerator)
 
 -- clash-protocols (me!)
 import Protocols
-import Protocols.Avalon.MemMap
-import Protocols.Avalon.Stream
-import Protocols.DfConv qualified as DfConv
+import Protocols.Experimental.Avalon.MemMap
+import Protocols.Experimental.Avalon.Stream
+import Protocols.Experimental.DfConv qualified as DfConv
 import Protocols.Hedgehog
 import Protocols.Internal
